:recycle: Basic refactoring to separate block and execution based feedback
