+++ runestone/hparsons/js/hparsons-sql.js
@@ -1,12 +1,8 @@
-import Handsontable from "handsontable";
 // import 'handsontable/dist/handsontable.full.css';
-import initSqlJs from "sql.js/dist/sql-wasm.js";
 import RunestoneBase from "../../common/js/runestonebase.js";
 import "../css/hparsons.css";
-import 'handsontable/dist/handsontable.full.css';
-import BlockBasedGrader from "./blockGrader.js";
-
-var allDburls = {};
+import BlockFeedback from "./BlockFeedback.js";
+import SQLFeedback from "./SQLFeedback.js";
 
 export var hpList;
 // Dictionary that contains all instances of horizontal Parsons problem objects
@@ -52,9 +48,20 @@ export default class SQLHParons extends RunestoneBase {
             this.suffix = this.code.substring(suffStart + 5);
             this.code = this.code.substring(0, suffStart);
         }
+
+        // Change to factory when more execution based feedback is included
+        if (this.isBlockGrading) {
+            this.feedbackController = new BlockFeedback(this);
+        } else {
+            this.feedbackController = new SQLFeedback(this);
+        }
+
+        // creating UI components
         this.createEditor();
         this.createOutput();
         this.createControls();
+        this.feedbackController.customizeUI();
+
         if ($(orig).data("caption")) {
             this.caption = $(orig).data("caption");
         } else {
@@ -63,191 +70,8 @@ export default class SQLHParons extends RunestoneBase {
         this.addCaption("runestone");
         this.indicate_component_ready();
 
-        // copied from activecode-sql
-        //  fnprefix sets the path to load the sql-wasm.wasm file
-        var bookprefix;
-        var fnprefix;
-        if (eBookConfig.useRunestoneServices) {
-            bookprefix = `${eBookConfig.app}/books/published/${eBookConfig.basecourse}`;
-            fnprefix = bookprefix + "/_static";
-        } else {
-            bookprefix = "";
-            fnprefix = "/_static";
-        }
-        this.config = {
-            locateFile: (filename) => `${fnprefix}/${filename}`,
-        };
-        var self = this;
-        initSqlJs(this.config).then(function (SQL) {
-            // set up call to load database asynchronously if given
-            if (self.dburl) {
-                if (self.dburl.startsWith("/_static")) {
-                    self.dburl = `${bookprefix}${self.dburl}`;
-                }
-                $(self.runButton).attr("disabled", "disabled");
-                let buttonText = $(self.runButton).text();
-                $(self.runButton).text($.i18n("msg_activecode_load_db"));
-                if (!(self.dburl in allDburls)) {
-                    allDburls[self.dburl] = {
-                        status: "loading",
-                        xWaitFor: jQuery.Deferred(),
-                    };
-                } else {
-                    if (allDburls[self.dburl].status == "loading") {
-                        allDburls[self.dburl].xWaitFor.done(function () {
-                            self.db = allDburls[self.dburl].dbObject;
-                            $(self.runButton).removeAttr("disabled");
-                            $(self.runButton).text(buttonText);
-                        });
-                        return;
-                    }
-                    self.db = allDburls[self.dburl].dbObject;
-                    $(self.runButton).removeAttr("disabled");
-                    $(self.runButton).text(buttonText);
-                    return;
-                }
-                var xhr = new XMLHttpRequest();
-                // For example: https://github.com/lerocha/chinook-database/raw/master/ChinookDatabase/DataSources/Chinook_Sqlite.sqlite
-                xhr.open("GET", self.dburl, true);
-                xhr.responseType = "arraybuffer";
-                xhr.onload = (e) => {
-                    var uInt8Array = new Uint8Array(xhr.response);
-                    self.db = new SQL.Database(uInt8Array);
-                    allDburls[self.dburl].dbObject = self.db;
-                    $(self.runButton).text(buttonText);
-                    $(self.runButton).removeAttr("disabled");
-                    allDburls[self.dburl].db = uInt8Array;
-                    allDburls[self.dburl].status = "ready";
-                    allDburls[self.dburl].xWaitFor.resolve();
-                    // contents is now [{columns:['col1','col2',...], values:[[first row], [second row], ...]}]
-                };
-                xhr.send();
-            } else {
-                self.db = new SQL.Database();
-            }
-        });
-
-        // Initializing options used for block based feedback
-        if (this.isBlockGrading) {
-            this.blockIndex = 0;
-            this.checkCount = 0;
-            this.numDistinct = 0;
-            this.hasSolved = false;
-            // TODO: not sure what is the best way to do this
-            this.grader = new BlockBasedGrader();
-            let solutionBlocks = [];
-            for (let i = 0; i < this.blockAnswer.length; ++i) {
-                solutionBlocks.push(this.originalBlocks[this.blockAnswer[i]]);
-            }
-            this.solution = solutionBlocks;
-            this.grader.solution = solutionBlocks;
-        }
-    }
-
-    // copied from activecode
-    async runButtonHandler() {
-        // Disable the run button until the run is finished.
-        this.runButton.disabled = true;
-        try {
-            await this.runProg();
-        } catch (e) {
-            console.log(`there was an error ${e} running the code`);
-        }
-        if (this.logResults) {
-            this.logCurrentAnswer();
-        }
-        this.renderFeedback();
-        // The run is finished; re-enable the button.
-        this.runButton.disabled = false;
-    }
-
-    // Called when check button clicked (block-based Feedback)
-    checkButtonHandler() {
-        this.checkCurrentAnswer();
-        this.renderFeedbackBlockBased();
-    }
-
-    // Used for block-based feedback
-    checkCurrentAnswer() {
-        if (!this.hasSolved) {
-            this.checkCount++;
-            this.clearFeedback();
-            this.grader.answer = this.hparsons.getParsonsTextArray();
-            this.grade = this.grader.grade();
-            if (this.grade == "correct") {
-                this.hasSolved = true;
-                this.correct = true;
-                $(this.checkButton).prop("disabled", true);
-            }
-        }
-    }
-
-    // Feedback UI for Block-based Feedback
-    clearFeedback() {
-        $(this.answerArea).removeClass("incorrect correct");
-        var children = this.answerArea.childNodes;
-        for (var i = 0; i < children.length; i++) {
-            $(children[i]).removeClass(
-                "correctPosition incorrectPosition"
-            );
-        }
-        $(this.messageDiv).hide();
-    }
-
-    renderFeedbackBlockBased() {
-        this.grade = this.grader.graderState;
-        var feedbackArea;
-        var answerArea = $(this.answerArea);
-        feedbackArea = $(this.messageDiv);
-
-        if (this.grade === "correct") {
-            answerArea.addClass("correct");
-            feedbackArea.fadeIn(100);
-            feedbackArea.attr("class", "alert alert-info");
-            if (this.checkCount > 1) {
-                feedbackArea.html(
-                    $.i18n("msg_parson_correct", this.checkCount)
-                );
-            } else {
-                feedbackArea.html($.i18n("msg_parson_correct_first_try"));
-            }
-        }
-
-        if (this.grade === "incorrectTooShort") {
-            // too little code
-            answerArea.addClass("incorrect");
-            feedbackArea.fadeIn(500);
-            feedbackArea.attr("class", "alert alert-danger");
-            feedbackArea.html($.i18n("msg_parson_too_short"));
-        }
-
-        if (this.grade === "incorrectMoveBlocks") {
-            var answerBlocks = this.answerArea.children;
-            var inSolution = [];
-            var inSolutionIndexes = [];
-            var notInSolution = [];
-            for (let i = 0; i < answerBlocks.length; i++) {
-                var block = answerBlocks[i];
-                var index = this.solution.indexOf(block.textContent);
-                if (index == -1) {
-                    notInSolution.push(block);
-                } else {
-                    inSolution.push(block);
-                    inSolutionIndexes.push(index);
-                }
-            }
-            var lisIndexes = this.grader.inverseLISIndices(inSolutionIndexes);
-            for (let i = 0; i < lisIndexes.length; i++) {
-                notInSolution.push(inSolution[lisIndexes[i]]);
-            }
-            answerArea.addClass("incorrect");
-            feedbackArea.fadeIn(500);
-            feedbackArea.attr("class", "alert alert-danger");
-            for (let i = 0; i < notInSolution.length; i++) {
-                $(notInSolution[i]).addClass("incorrectPosition");
-            }
-            feedbackArea.html($.i18n("msg_parson_wrong_order"));
-        }
+        // initializing functionalities for different feedback
+        this.feedbackController.init();
     }
 
 
@@ -271,9 +95,7 @@ export default class SQLHParons extends RunestoneBase {
         this.outerDiv.innerHTML = parsonsHTML;
         this.outerDiv.addEventListener('horizontal-parsons', (ev) => {
             this.logHorizontalParsonsEvent(ev.detail)
-            if (this.isBlockGrading) {
-                this.clearFeedback();
-            }
+            this.feedbackController.clearFeedback();
         })
         let blocks = [];
         let blockIndex = this.code.indexOf('--blocks--');
@@ -283,65 +105,32 @@ export default class SQLHParons extends RunestoneBase {
             blocksString = endIndex > -1 ? blocksString.substring(0, endIndex) : blocksString;
             blocks = blocksString.split('\n');
         }
-        this.hparsons = $(this.outerDiv).find("horizontal-parsons")[0];
+        this.hparsonsInput = $(this.outerDiv).find("horizontal-parsons")[0];
         this.originalBlocks = blocks.slice(1,-1);
-        this.hparsons.parsonsData = blocks.slice(1,-1);
+        this.hparsonsInput.parsonsData = blocks.slice(1,-1);
 
         // Saving UI area for block-based feedback
-        this.answerArea = this.hparsons.shadowRoot.querySelector('.drop-area');
+        this.answerArea = this.hparsonsInput.shadowRoot.querySelector('.drop-area');
     }
 
     createOutput() {
-        if (this.isBlockGrading) {
-            // Block based grading output
-            this.messageDiv = document.createElement("div");
-            this.outerDiv.appendChild(this.messageDiv);
-        } else {
-            // Execution based grading output
-            // Create a parent div with two elements:  pre for standard output and a div
-            // to hold turtle graphics output.  We use a div in case the turtle changes from
-            var outDiv = document.createElement("div");
-            $(outDiv).addClass("hp_output col-md-12");
-            this.outDiv = outDiv;
-            this.output = document.createElement("pre");
-            this.output.id = this.divid + "_stdout";
-            $(this.output).css("visibility", "hidden");
-            var clearDiv = document.createElement("div");
-            $(clearDiv).css("clear", "both"); // needed to make parent div resize properly
-            this.outerDiv.appendChild(clearDiv);
-            outDiv.appendChild(this.output);
-            this.outerDiv.appendChild(outDiv);
-            clearDiv = document.createElement("div");
-            $(clearDiv).css("clear", "both"); // needed to make parent div resize properly
-            this.outerDiv.appendChild(clearDiv);
-        }
+        this.feedbackController.createOutput();
     }
 
     // copied from activecode
     createControls() {
         var ctrlDiv = document.createElement("div");
-        var butt;
         $(ctrlDiv).addClass("hp_actions");
         $(ctrlDiv).addClass("col-md-12");
 
-        if (this.isBlockGrading) {
-            // Check button: if line based feedback
-            butt = document.createElement("button");
-            $(butt).text("Check");
-            $(butt).addClass("btn btn-success run-button");
-            ctrlDiv.appendChild(butt);
-            this.checkButton = butt;
-            this.checkButton.onclick = this.checkButtonHandler.bind(this);
-            $(butt).attr("type", "button");
-        } else {
-            // Run button: if execution based feedback
-            butt = document.createElement("button");
-            $(butt).text("Run");
-            $(butt).addClass("btn btn-success run-button");
-            ctrlDiv.appendChild(butt);
-            this.runButton = butt;
-            this.runButton.onclick = this.runButtonHandler.bind(this);
-            $(butt).attr("type", "button");
+        // Run Button
+        this.runButton = document.createElement("button");
+        $(this.runButton).addClass("btn btn-success run-button");
+        ctrlDiv.appendChild(this.runButton);
+        $(this.runButton).attr("type", "button");
+        $(this.runButton).text("Run");
+        this.runButton.onclick = () => {
+            this.feedbackController.runButtonHandler();
         }
 
         // Reset button
@@ -352,213 +141,13 @@ export default class SQLHParons extends RunestoneBase {
         ctrlDiv.appendChild(resetBtn);
         this.resetButton = resetBtn;
         this.resetButton.onclick = () => {
-            this.hparsons.resetInput();
-            if (this.isBlockGrading) {
-                this.clearFeedback();
-                $(this.checkButton).prop("disabled", false);
-                this.checkCount = 0;
-                this.hasSolved = false;
-            }
+            this.hparsonsInput.resetInput();
+            this.feedbackController.clearFeedback();
         }
         $(resetBtn).attr("type", "button");
 
         $(this.outerDiv).prepend(ctrlDiv);
         this.controlDiv = ctrlDiv;
-    }
-
-    // copied from activecode-sql
-    async runProg(noUI, logResults) {
-        if (typeof logResults === "undefined") {
-            this.logResults = true;
-        } else {
-            this.logResults = logResults;
-        }
-        if (typeof noUI !== "boolean") {
-            noUI = false;
-        }
-        // Clear any old results
-        this.saveCode = "True";
-        let divid = this.divid + "_sql_out";
-        let respDiv = document.getElementById(divid);
-        if (respDiv) {
-            respDiv.parentElement.removeChild(respDiv);
-        }
-        $(this.output).text("");
-        // Run this query
-        let query = await this.buildProg(false); // false --> Do not include suffix
-        if (!this.db) {
-            $(this.output).text(
-                `Error: Database not initialized! DBURL: ${this.dburl}`
-            );
-            return;
-        }
-
-        let it = this.db.iterateStatements(query);
-        this.results = [];
-        try {
-            for (let statement of it) {
-                let columns = statement.getColumnNames();
-                if (columns.length > 0) {
-                    // data! probably a SELECT
-                    let data = [];
-                    while (statement.step()) {
-                        data.push(statement.get());
-                    }
-                    this.results.push({
-                        status: "success",
-                        columns: columns,
-                        values: data,
-                        rowcount: data.length,
-                    });
-                } else {
-                    let nsql = statement.getNormalizedSQL();
-                    let prefix = nsql.substr(0, 6).toLowerCase();
-                    statement.step(); // execute the query
-                    // Try to detect INSERT/UPDATE/DELETE to give friendly feedback
-                    // on rows modified - unfortunately, this won't catch such queries
-                    // if they use CTEs.  There seems to be no reliable way of knowing
-                    // when a SQLite query actually modified data.
-                    if (
-                        prefix === "insert" ||
-                        prefix === "update" ||
-                        prefix === "delete"
-                    ) {
-                        this.results.push({
-                            status: "success",
-                            operation: prefix,
-                            rowcount: this.db.getRowsModified(),
-                        });
-                    } else {
-                        this.results.push({ status: "success" });
-                    }
-                }
-            }
-        } catch (e) {
-            this.results.push({
-                status: "failure",
-                message: e.toString(),
-                sql: it.getRemainingSQL(),
-            });
-        }
-
-        if (this.results.length === 0) {
-            this.results.push({
-                status: "failure",
-                message: "No queries submitted.",
-            });
-        }
-
-        respDiv = document.createElement("div");
-        respDiv.id = divid;
-        this.outDiv.appendChild(respDiv);
-        $(this.outDiv).show();
-        // Sometimes we don't want to show a bunch of intermediate results
-        // like when we are including a bunch of previous statements from
-        // other activecodes In that case the showlastsql flag can be set
-        // so we only show the last result
-        let resultArray = this.results;
-        for (let r of resultArray) {
-            let section = document.createElement("div");
-            section.setAttribute("class", "hp_sql_result");
-            respDiv.appendChild(section);
-            if (r.status === "success") {
-                if (r.columns) {
-                    let tableDiv = document.createElement("div");
-                    section.appendChild(tableDiv);
-                    let maxHeight = 350;
-                    if (resultArray.length > 1) maxHeight = 200; // max height smaller if lots of results
-                    createTable(r, tableDiv, maxHeight);
-                    let messageBox = document.createElement("pre");
-                    let rmsg = r.rowcount !== 1 ? " rows " : " row ";
-                    let msg = "" + r.rowcount + rmsg + "returned";
-                    if (r.rowcount > 100) {
-                        msg = msg + " (only first 100 rows displayed)";
-                    }
-                    msg = msg + ".";
-                    messageBox.textContent = msg;
-                    messageBox.setAttribute("class", "hp_sql_result_success");
-                    section.appendChild(messageBox);
-                } else if (r.rowcount) {
-                    let messageBox = document.createElement("pre");
-                    let op = r.operation;
-                    op = op + (op.charAt(op.length - 1) === "e" ? "d." : "ed.");
-                    let rmsg = r.rowcount !== 1 ? " rows " : " row ";
-                    messageBox.textContent = "" + r.rowcount + rmsg + op;
-                    messageBox.setAttribute("class", "hp_sql_result_success");
-                    section.appendChild(messageBox);
-                } else {
-                    let messageBox = document.createElement("pre");
-                    messageBox.textContent = "Operation succeeded.";
-                    messageBox.setAttribute("class", "hp_sql_result_success");
-                    section.appendChild(messageBox);
-                }
-            } else {
-                let messageBox = document.createElement("pre");
-                messageBox.textContent = r.message;
-                messageBox.setAttribute("class", "hp_sql_result_failure");
-                section.appendChild(messageBox);
-            }
-        }
-
-        // Now handle autograding
-        if (this.suffix) {
-            this.testResult = this.autograde(
-                this.results[this.results.length - 1]
-            );
-        } else {
-            $(this.output).css("visibility", "hidden");
-        }
-
-        return Promise.resolve("done");
-    }
-
-    // copied from anctivecode
-    // changed to getting parsons
-    async buildProg(useSuffix) {
-        // assemble code from prefix, suffix, and editor for running.
-        var prog;
-        if (this.textentry) {
-            prog = this.hparsons.getCurrentInput();
-        } else {
-            prog = this.hparsons.getParsonsTextArray().join(' ') + "\n";
-        }
-        this.pretext = "";
-        this.pretextLines = 0;
-        this.progLines = prog.match(/\n/g).length + 1;
-        if (useSuffix && this.suffix) {
-            prog = prog + this.suffix;
-        }
-        return Promise.resolve(prog);
-    }
-
-    // copied from activecode-sql
-    async logCurrentAnswer(sid) {
-        let data = {
-            div_id: this.divid,
-            code: this.hparsons.getParsonsTextArray(),
-            language: "sql",
-            // errinfo: this.results[this.results.length - 1].status,
-            to_save: this.saveCode,
-            prefix: this.pretext,
-            suffix: this.suffix,
-        }; // Log the run event
-        if (typeof sid !== "undefined") {
-            data.sid = sid;
-        }
-        await this.logRunEvent(data);
-
-        if (this.unit_results) {
-            let unitData = {
-                event: "unittest",
-                div_id: this.divid,
-                course: eBookConfig.course,
-                act: this.unit_results,
-            };
-            if (typeof sid !== "undefined") {
-                unitData.sid = sid;
-            }
-            await this.logBookEvent(unitData);
-        }
     }
 
     logHorizontalParsonsEvent(hparsonsEvent) {
@@ -571,127 +160,6 @@ export default class SQLHParons extends RunestoneBase {
         };
         this.logBookEvent(ev);
     }
-
-    renderFeedback() {
-        if (this.testResult) {
-            $(this.output).text(this.testResult);
-            $(this.output).css("visibility", "visible");
-        }
-    }
-
-    autograde(result_table) {
-        var tests = this.suffix.split(/\n/);
-        this.passed = 0;
-        this.failed = 0;
-        // Tests should be of the form
-        // assert row,col oper value for example
-        // assert 4,4 == 3
-        var result = "";
-        tests = tests.filter(function (s) {
-            return s.indexOf("assert") > -1;
-        });
-        for (let test of tests) {
-            let wlist = test.split(/\s+/);
-            wlist.shift();
-            let loc = wlist.shift();
-            let oper = wlist.shift();
-            let expected = wlist.join(" ");
-            let [row, col] = loc.split(",");
-            result += this.testOneAssert(
-                row,
-                col,
-                oper,
-                expected,
-                result_table
-            );
-            result += "\n";
-        }
-        let pct = (100 * this.passed) / (this.passed + this.failed);
-        pct = pct.toLocaleString(undefined, { maximumFractionDigits: 2 });
-        result += `You passed ${this.passed} out of ${this.passed + this.failed
-            } tests for ${pct}%`;
-        this.unit_results = `percent:${pct}:passed:${this.passed}:failed:${this.failed}`;
-        return result;
-    }
-    testOneAssert(row, col, oper, expected, result_table) {
-        // make sure row and col are in bounds
-        let actual;
-        let output = "";
-        try {
-            actual = result_table.values[row][col];
-        } catch (e) {
-            if (expected == 'NO_DATA') {
-                this.passed++;
-                output = `Passed: No data in row ${row}, column ${col}`;
-                return output;
-            } else {
-                output = `Failed: Not enough data to check row ${row} or column ${col}`;
-                return output;
-            }
-        }
-        const operators = {
-            "==": function (operand1, operand2) {
-                return operand1 == operand2;
-            },
-            "!=": function (operand1, operand2) {
-                return operand1 != operand2;
-            },
-            ">": function (operand1, operand2) {
-                return operand1 > operand2;
-            },
-            "<": function (operand1, operand2) {
-                return operand1 > operand2;
-            },
-        };
-        let res = operators[oper](actual, expected);
-        if (res) {
-            output = `Pass: ${actual} ${oper} ${expected} in row ${row} column ${result_table.columns[col]}`;
-            this.passed++;
-        } else {
-            output = `Failed ${actual} ${oper} ${expected} in row ${row} column ${result_table.columns[col]}`;
-            this.failed++;
-        }
-        return output;
-    }
-}
-
-function createTable(tableData, container, maxHeight) {
-    let data = tableData.values;
-    let trimRows = undefined;
-    if (data.length === 0) {
-        // kludge: no column headers will show up unless we do this
-        data = [tableData.columns.map((e) => null)];
-        trimRows = [0];
-    }
-
-    var hot = new Handsontable(container, {
-        data: data,
-        trimRows: trimRows,
-        width: "100%",
-        height: maxHeight,
-        autoRowSize: true,
-        autoColumnSize: { useHeaders: true },
-        rowHeaders: false,
-        colHeaders: tableData.columns,
-        editor: false,
-        maxRows: 100,
-        filters: false,
-        dropdownMenu: false,
-        licenseKey: "non-commercial-and-evaluation",
-    });
-
-    // calculate actual height and resize
-    let actualHeight = 40; // header height + small margin
-    if (tableData.values.length > 0) {
-        for (let i = 0; i < data.length; i++) {
-            actualHeight = actualHeight + hot.getRowHeight(i);
-            if (actualHeight > maxHeight) break;
-        }
-    }
-
-    hot.updateSettings({ height: actualHeight });
-
-    return hot;
 }
